refactor: 更新数据模型和数据库访问层，调整主键类型为 Integer，优化表结构

+++ avalon-application/avalon-admin-server/src/main/java/io/github/lishangbu/avalon/admin/controller/dataset/BerryController.java
@@ -70,7 +70,7 @@ public class BerryController {
    * @param id Berry 主键
    */
   @DeleteMapping("/{id:\\d+}")
-  public void deleteById(@PathVariable Long id) {
+  public void deleteById(@PathVariable Integer id) {
     berryService.removeById(id);
   }
 }
